+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.21
 - update to bevy `0.12` (@raffaeleragni, https://github.com/jakobhellermann/bevy-inspector-egui/pull/157)
 - add ability to edit and insert elements into maps (@B-Reif, https://github.com/jakobhellermann/bevy-inspector-egui/pull/155)
+- show asset path if available in inspector
 
 ## Version 0.20
 - update egui to `0.23`